Alert support
- Support a 3rdtype of query, "Alerts" to go with "Basic" and "CAQL"
- Support alert queries via annotationQuery to draw alerts over graphs

Both of these support querying by a fixed alert_id *or* via the standard
circonus search API for alerts.  In addition, both queries support using
a local filter on data retrieved via the API to further reduce the queries
by things not supported in the circonus API (tags for example).